Remove static portfolio and certifications multilanguage content
- Move out Portfolio and Certifications type definitions under actions/contentful folder.
