+++ src/actions/contentful/get-certifications.ts
@@ -24,7 +24,7 @@ const ACCESS_TOKEN = process.env.API_CONTENTFUL_ACCESS_TOKEN ?? '';
 // eslint-disable-next-line max-len
 const URL = `https://graphql.contentful.com/content/v1/spaces/${SPACE_ID}/environments/master`;
 
-export default async (): Promise<CertificationNew[]> => {
+export default async (): Promise<Certification[]> => {
   const fetchResponse = await fetch(URL, {
     next: { revalidate: 3600 },
     method: 'POST',
@@ -45,7 +45,7 @@ export default async (): Promise<CertificationNew[]> => {
       CertificationContentful
     >;
 
-  return jsonResponse.data.certificationCollection.items.map<CertificationNew>(
+  return jsonResponse.data.certificationCollection.items.map<Certification>(
     ({ issued, certificationIssuer, ...certification }) => ({
       ...certification,
       issued: new Date(issued),
