+++ src/actions/contentful/types.certification.d.ts
@@ -5,15 +5,5 @@ type Certification = {
   issuedBy: Tech;
   certificateLink: string;
   courseLink: string;
-  description: string;
-};
-
-type CertificationNew = {
-  id: string;
-  title: string;
-  issued: Date;
-  issuedBy: Tech;
-  certificateLink: string;
-  courseLink: string;
   description: MultiLanguageData<string>;
 };
