--- src/languages/certifications/en.ts
@@ -1,51 +1,0 @@
-/* eslint-disable max-len */
-
-const certification: Certification[] = [
-  {
-    id: '2XU86S65RU75',
-    title: 'Functional Programming Principles in Scala',
-    issued: new Date('2018-01-01'),
-    issuedBy: 'Coursera',
-    certificateLink:
-      'https://www.coursera.org/account/accomplishments/verify/2XU86S65RU75',
-    courseLink: 'https://www.coursera.org/learn/scala-functional-programming',
-    description:
-      'A course provided by Bonzzu as an educational program to increase our knowledge and skills in the Scala programming language for the Jemstep project. This course was instructed by the designer of the language, Martin Odersky.',
-  },
-  {
-    id: 'W94PPTA39266',
-    title: 'Functional Program Design in Scala',
-    issued: new Date('2018-02-01'),
-    issuedBy: 'Coursera',
-    certificateLink:
-      'https://www.coursera.org/account/accomplishments/verify/W94PPTA39266',
-    courseLink:
-      'https://www.coursera.org/learn/scala-functional-program-design',
-    description:
-      'A course provided by Bonzzu as an educational program to increase our knowledge and skills in the Scala programming language for the Jemstep project. This course was instructed by the designer of the language, Martin Odersky.',
-  },
-  {
-    id: 'UC-3dc1a04f-09e0-4992-8076-2c1eec6e1f39',
-    title: 'Understanding Typescript',
-    issued: new Date('2023-02-01'),
-    issuedBy: 'Udemy',
-    certificateLink:
-      'https://www.udemy.com/certificate/UC-3dc1a04f-09e0-4992-8076-2c1eec6e1f39/',
-    courseLink:
-      'https://www.udemy.com/course/understanding-typescript/?couponCode=OF83024E',
-    description: 'I took this course to improve my skills in TypeScript.',
-  },
-  {
-    id: 'UC-c396fa71-c367-457c-8d88-c077b105a706',
-    title: 'Node.js, Express, MongoDB & More: The Complete Bootcamp 2023',
-    issued: new Date('2023-06-01'),
-    issuedBy: 'Udemy',
-    certificateLink:
-      'https://www.udemy.com/certificate/UC-c396fa71-c367-457c-8d88-c077b105a706/',
-    courseLink: 'https://www.udemy.com/course/nodejs-express-mongodb-bootcamp',
-    description:
-      'Back in my college years, I came across Node.js. Running JavaScript on the server side? That caught my attention, and I started learning that JavaScript runtime environment. In 2023, I decided to take this course to refresh my skills in Node.js.',
-  },
-];
-
-export default certification;
