--- src/languages/certifications/es.ts
@@ -1,51 +1,0 @@
-/* eslint-disable max-len */
-
-const certification: Certification[] = [
-  {
-    id: '2XU86S65RU75',
-    title: 'Functional Programming Principles in Scala',
-    issued: new Date('2018-01-01'),
-    issuedBy: 'Coursera',
-    certificateLink:
-      'https://www.coursera.org/account/accomplishments/verify/2XU86S65RU75',
-    courseLink: 'https://www.coursera.org/learn/scala-functional-programming',
-    description:
-      'Un curso ofrecido por Bonzzu como un programa educativo para aumentar nuestros conocimientos y habilidades en el lenguaje de programación Scala para el proyecto Jemstep. Este curso fue impartido por el diseñador del lenguaje, Martin Odersky.',
-  },
-  {
-    id: 'W94PPTA39266',
-    title: 'Functional Program Design in Scala',
-    issued: new Date('2018-02-01'),
-    issuedBy: 'Coursera',
-    certificateLink:
-      'https://www.coursera.org/account/accomplishments/verify/W94PPTA39266',
-    courseLink:
-      'https://www.coursera.org/learn/scala-functional-program-design',
-    description:
-      'Un curso ofrecido por Bonzzu como un programa educativo para aumentar nuestros conocimientos y habilidades en el lenguaje de programación Scala para el proyecto Jemstep. Este curso fue impartido por el diseñador del lenguaje, Martin Odersky.',
-  },
-  {
-    id: 'UC-3dc1a04f-09e0-4992-8076-2c1eec6e1f39',
-    title: 'Understanding Typescript',
-    issued: new Date('2023-02-01'),
-    issuedBy: 'Udemy',
-    certificateLink:
-      'https://www.udemy.com/certificate/UC-3dc1a04f-09e0-4992-8076-2c1eec6e1f39/',
-    courseLink:
-      'https://www.udemy.com/course/understanding-typescript/?couponCode=OF83024E',
-    description: 'Tomé este curso para mejorar mis habilidades en TypeScript.',
-  },
-  {
-    id: 'UC-c396fa71-c367-457c-8d88-c077b105a706',
-    title: 'Node.js, Express, MongoDB & More: The Complete Bootcamp 2023',
-    issued: new Date('2023-06-01'),
-    issuedBy: 'Udemy',
-    certificateLink:
-      'https://www.udemy.com/certificate/UC-c396fa71-c367-457c-8d88-c077b105a706/',
-    courseLink: 'https://www.udemy.com/course/nodejs-express-mongodb-bootcamp',
-    description:
-      'Durante mis años universitarios, me encontré con Node.js. ¿Correr JavaScript en el lado del servidor? Eso llamó mi atención, y comencé a aprender sobre ese entorno de ejecución de JavaScript. En 2023, decidí tomar este curso para refrescar mis habilidades en Node.js.',
-  },
-];
-
-export default certification;
