--- src/languages/portfolio/index.ts
@@ -1,7 +1,0 @@
-import en from './en';
-import es from './es';
-
-const verbiage = { en, es };
-
-export default verbiage;
-export { en, es };
